refactor: change script structure
