+++ tasks/1-EDA/test_eda/local_config.yml
@@ -1,0 +1,9 @@
+paths:
+  data_folder: ../../../data/rsna-miccai-brain-tumor-radiogenomic-classification
+
+eda:
+
+general:
+
+training:
+  prop: 0.5
